ceph-dencoder: parse ull args

+++ src/test/encoding/ceph_dencoder.cc
@@ -370,7 +370,7 @@ int main(int argc, const char **argv)
 	usage(cerr);
 	exit(1);
       }
-      features = atoi(*i);
+      features = atoll(*i);
     } else if (*i == string("encode")) {
       if (!den) {
 	cerr << "must first select type with 'type <name>'" << std::endl;
